[integration] Create M3Resources interface with current code as Docker-backed impl
This commit takes the existing implementation of Node
and Coordinator and moves it to a docker package where it will
exist as a Docker-backed implementation of those interfaces.
Additionally, rename DockerResources to M3Resources. These changes
give us a vehicle for providing an in-process backed implementation
of M3Resources, Node, and Coordinator.
