+++ src/integration/resources/resources.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2020 Uber Technologies, Inc.
+// Copyright (c) 2021  Uber Technologies, Inc.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -18,24 +18,20 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-// Package resources contains resources needed to setup docker containers for M3 tests.
 package resources
 
 import (
 	"time"
 
+	protobuftypes "github.com/gogo/protobuf/types"
+	"go.uber.org/zap"
+
 	"github.com/m3db/m3/src/dbnode/generated/proto/namespace"
 	"github.com/m3db/m3/src/query/generated/proto/admin"
-	xerrors "github.com/m3db/m3/src/x/errors"
 	"github.com/m3db/m3/src/x/instrument"
-
-	protobuftypes "github.com/gogo/protobuf/types"
-	"github.com/ory/dockertest/v3"
-	"go.uber.org/zap"
 )
 
 const (
-	timeout   = time.Second * 60
 	retention = "6h"
 
 	// AggName is the name of the aggregated namespace.
@@ -46,155 +42,8 @@ const (
 	ColdWriteNsName = "coldWritesRepairAndNoIndex"
 )
 
-// DockerResources represents a set of dockerized test components.
-type DockerResources interface {
-	// Cleanup closes and removes all corresponding containers.
-	Cleanup() error
-	// Nodes returns all node resources.
-	Nodes() Nodes
-	// Coordinator returns the coordinator resource.
-	Coordinator() Coordinator
-}
-
-// ClusterOptions represents a set of options for a cluster setup.
-type ClusterOptions struct {
-	ReplicationFactor int32
-	NumShards         int32
-}
-
-type dockerResources struct {
-	coordinator Coordinator
-	nodes       Nodes
-
-	pool *dockertest.Pool
-}
-
-// SetupSingleM3DBNode creates docker resources representing a setup with a
-// single DB node.
-func SetupSingleM3DBNode(opts ...SetupOptions) (DockerResources, error) { // nolint: gocyclo
-	options := setupOptions{}
-	for _, f := range opts {
-		f(&options)
-	}
-
-	pool, err := dockertest.NewPool("")
-	if err != nil {
-		return nil, err
-	}
-
-	pool.MaxWait = timeout
-
-	if !options.existingCluster {
-		if err := setupNetwork(pool); err != nil {
-			return nil, err
-		}
-
-		if err := setupVolume(pool); err != nil {
-			return nil, err
-		}
-	}
-
-	iOpts := instrument.NewOptions()
-	dbNode, err := newDockerHTTPNode(pool, dockerResourceOptions{
-		image:         options.dbNodeImage,
-		containerName: options.dbNodeContainerName,
-		iOpts:         iOpts,
-	})
-
-	success := false
-	dbNodes := Nodes{dbNode}
-	defer func() {
-		// NB: only defer close in the failure case, otherwise calling function
-		// is responsible for closing the resources.
-		if !success {
-			for _, dbNode := range dbNodes {
-				if dbNode != nil {
-					dbNode.Close()
-				}
-			}
-		}
-	}()
-
-	if err != nil {
-		return nil, err
-	}
-
-	coordinator, err := newDockerHTTPCoordinator(pool, dockerResourceOptions{
-		image:         options.coordinatorImage,
-		containerName: options.coordinatorContainerName,
-		iOpts:         iOpts,
-	})
-
-	defer func() {
-		// NB: only defer close in the failure case, otherwise calling function
-		// is responsible for closing the resources.
-		if !success && coordinator != nil {
-			coordinator.Close()
-		}
-	}()
-
-	if err != nil {
-		return nil, err
-	}
-
-	cluster := &dockerResources{
-		coordinator: coordinator,
-		nodes:       dbNodes,
-		pool:        pool,
-	}
-	err = SetupCluster(cluster, nil)
-
-	logger := iOpts.Logger().With(zap.String("source", "harness"))
-	logger.Info("all healthy")
-	success = true
-	return cluster, err
-}
-
-// AttachToExistingContainers attaches docker API to an existing coordinator
-// and one or more dbnode containers.
-func AttachToExistingContainers(
-	coordinatorContainerName string,
-	dbNodesContainersNames []string,
-) (DockerResources, error) {
-	pool, err := dockertest.NewPool("")
-	if err != nil {
-		return nil, err
-	}
-	pool.MaxWait = timeout
-
-	iOpts := instrument.NewOptions()
-	dbNodes := Nodes{}
-	for _, containerName := range dbNodesContainersNames {
-		dbNode, err := newDockerHTTPNode(pool, dockerResourceOptions{
-			iOpts:         iOpts,
-			containerName: containerName,
-		})
-		if err != nil {
-			return nil, err
-		}
-		dbNodes = append(dbNodes, dbNode)
-	}
-
-	coordinator, err := newDockerHTTPCoordinator(
-		pool,
-		dockerResourceOptions{
-			iOpts:         iOpts,
-			containerName: coordinatorContainerName,
-		},
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	return &dockerResources{
-		coordinator: coordinator,
-		nodes:       dbNodes,
-		pool:        pool,
-	}, err
-}
-
-// SetupCluster setupts m3 cluster on provided docker containers.
-func SetupCluster(cluster DockerResources, opts *ClusterOptions) error { // nolint: gocyclo
+// SetupCluster setups m3 cluster on provided docker containers.
+func SetupCluster(cluster M3Resources, opts *ClusterOptions) error { // nolint: gocyclo
 	coordinator := cluster.Coordinator()
 	iOpts := instrument.NewOptions()
 	logger := iOpts.Logger().With(zap.String("source", "harness"))
@@ -301,7 +150,7 @@ func SetupCluster(cluster DockerResources, opts *ClusterOptions) error { // noli
 	}
 
 	logger.Info("waiting for healthy")
-	if err := cluster.Nodes().waitForHealthy(); err != nil {
+	if err := cluster.Nodes().WaitForHealthy(); err != nil {
 		return err
 	}
 
@@ -313,25 +162,3 @@ func SetupCluster(cluster DockerResources, opts *ClusterOptions) error { // noli
 	logger.Info("all healthy")
 	return nil
 }
-
-func (r *dockerResources) Cleanup() error {
-	if r == nil {
-		return nil
-	}
-
-	var multiErr xerrors.MultiError
-	if r.coordinator != nil {
-		multiErr = multiErr.Add(r.coordinator.Close())
-	}
-
-	for _, dbNode := range r.nodes {
-		if dbNode != nil {
-			multiErr = multiErr.Add(dbNode.Close())
-		}
-	}
-
-	return multiErr.FinalError()
-}
-
-func (r *dockerResources) Nodes() Nodes             { return r.nodes }
-func (r *dockerResources) Coordinator() Coordinator { return r.coordinator }
